+++ opensrp-app/src/main/java/org/smartregister/sync/intent/SyncIntentService.java
@@ -1,435 +1,25 @@
 package org.smartregister.sync.intent;
 
-import android.content.Context;
 import android.content.Intent;
-import android.util.Pair;
 
-import org.apache.commons.lang3.StringUtils;
-import org.joda.time.DateTime;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-import org.smartregister.AllConstants;
-import org.smartregister.CoreLibrary;
-import org.smartregister.R;
-import org.smartregister.SyncConfiguration;
-import org.smartregister.domain.FetchStatus;
-import org.smartregister.domain.Response;
-import org.smartregister.domain.Task;
-import org.smartregister.domain.db.EventClient;
-import org.smartregister.receiver.SyncStatusBroadcastReceiver;
-import org.smartregister.repository.EventClientRepository;
-import org.smartregister.service.HTTPAgent;
-import org.smartregister.sync.helper.ECSyncHelper;
-import org.smartregister.util.NetworkUtils;
-import org.smartregister.util.SyncUtils;
-import org.smartregister.view.activity.DrishtiApplication;
-
-import java.text.MessageFormat;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
-import java.util.Map;
-
-import timber.log.Timber;
+import org.smartregister.sync.helper.SyncServiceHelper;
 
 public class SyncIntentService extends BaseSyncIntentService {
-    private static final String ADD_URL = "/rest/event/add";
-    public static final String SYNC_URL = "/rest/event/sync";
-
-    private Context context;
-    private HTTPAgent httpAgent;
-    private SyncUtils syncUtils;
-
-    protected static final int EVENT_PULL_LIMIT = 250;
-    protected static final int EVENT_PUSH_LIMIT = 50;
-
     public SyncIntentService() {
         super("SyncIntentService");
     }
 
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
-        context = getBaseContext();
-        httpAgent = CoreLibrary.getInstance().context().getHttpAgent();
-        syncUtils = new SyncUtils(getBaseContext());
         return super.onStartCommand(intent, flags, startId);
     }
 
     @Override
     protected void onHandleIntent(Intent intent) {
         super.onHandleIntent(intent);
-        handleSync();
-    }
-
-    protected void handleSync() {
-        sendSyncStatusBroadcastMessage(FetchStatus.fetchStarted);
-
-        doSync();
-    }
-
-    private void doSync() {
-        if (!NetworkUtils.isNetworkAvailable()) {
-            complete(FetchStatus.noConnection);
-            return;
-        }
-
-        try {
-            boolean hasValidAuthorization = syncUtils.verifyAuthorization();
-            if (hasValidAuthorization || !CoreLibrary.getInstance().getSyncConfiguration().disableSyncToServerIfUserIsDisabled()) {
-                pushToServer();
-            }
-            if (hasValidAuthorization) {
-                pullECFromServer();
-            } else {
-                syncUtils.logoutUser();
-            }
-
-        } catch (Exception e) {
-            Timber.e(e);
-            complete(FetchStatus.fetchedFailed);
-        }
+        SyncServiceHelper syncServiceHelper = SyncServiceHelper.getInstance();
+        syncServiceHelper.handleSync();
     }
 
 
-    private void pullECFromServer() {
-        fetchRetry(0);
-    }
-
-    private synchronized void fetchRetry(final int count) {
-        try {
-            SyncConfiguration configs = CoreLibrary.getInstance().getSyncConfiguration();
-            if (configs.getSyncFilterParam() == null || StringUtils.isBlank(configs.getSyncFilterValue())) {
-                complete(FetchStatus.fetchedFailed);
-                return;
-            }
-
-            final ECSyncHelper ecSyncUpdater = ECSyncHelper.getInstance(context);
-            String baseUrl = CoreLibrary.getInstance().context().
-                    configuration().dristhiBaseURL();
-            if (baseUrl.endsWith("/")) {
-                baseUrl = baseUrl.substring(0, baseUrl.lastIndexOf("/"));
-            }
-
-            Long lastSyncDatetime = ecSyncUpdater.getLastSyncTimeStamp();
-            Timber.i("LAST SYNC DT %s", new DateTime(lastSyncDatetime));
-
-            if (httpAgent == null) {
-                complete(FetchStatus.fetchedFailed);
-            }
-
-            String url = baseUrl + SYNC_URL;
-            Response resp;
-            if (configs.isSyncUsingPost()) {
-                JSONObject syncParams = new JSONObject();
-                syncParams.put(configs.getSyncFilterParam().value(), configs.getSyncFilterValue());
-                syncParams.put("serverVersion", lastSyncDatetime);
-                syncParams.put("limit", getEventPullLimit());
-                resp = httpAgent.postWithJsonResponse(url, syncParams.toString());
-            } else {
-                url += "?" + configs.getSyncFilterParam().value() + "=" + configs.getSyncFilterValue() + "&serverVersion=" + lastSyncDatetime + "&limit=" + getEventPullLimit();
-                Timber.i("URL: %s", url);
-                resp = httpAgent.fetch(url);
-            }
-
-            if (resp.isUrlError()) {
-                FetchStatus.fetchedFailed.setDisplayValue(resp.status().displayValue());
-                complete(FetchStatus.fetchedFailed);
-                return;
-            }
-
-            if (resp.isTimeoutError()) {
-                FetchStatus.fetchedFailed.setDisplayValue(resp.status().displayValue());
-                complete(FetchStatus.fetchedFailed);
-            }
-
-            if (resp.isFailure() && !resp.isUrlError() && !resp.isTimeoutError()) {
-                fetchFailed(count);
-            }
-            int eCount;
-            JSONObject jsonObject = new JSONObject();
-            if (resp.payload() == null) {
-                eCount = 0;
-            } else {
-                jsonObject = new JSONObject((String) resp.payload());
-                eCount = fetchNumberOfEvents(jsonObject);
-                Timber.i("Parse Network Event Count: %s", eCount);
-            }
-
-            if (eCount == 0) {
-                complete(FetchStatus.nothingFetched);
-            } else if (eCount < 0) {
-                fetchFailed(count);
-            } else {
-                final Pair<Long, Long> serverVersionPair = getMinMaxServerVersions(jsonObject);
-                long lastServerVersion = serverVersionPair.second - 1;
-                if (eCount < getEventPullLimit()) {
-                    lastServerVersion = serverVersionPair.second;
-                }
-
-                boolean isSaved = ecSyncUpdater.saveAllClientsAndEvents(jsonObject);
-                //update sync time if all event client is save.
-                if (isSaved) {
-                    processClient(serverVersionPair);
-                    ecSyncUpdater.updateLastSyncTimeStamp(lastServerVersion);
-                }
-                fetchRetry(0);
-            }
-        } catch (Exception e) {
-            Timber.e(e, "Fetch Retry Exception:  %s", e.getMessage());
-            fetchFailed(count);
-        }
-    }
-
-    public void fetchFailed(int count) {
-        if (count < CoreLibrary.getInstance().getSyncConfiguration().getSyncMaxRetries()) {
-            int newCount = count + 1;
-            fetchRetry(newCount);
-        } else {
-            complete(FetchStatus.fetchedFailed);
-        }
-    }
-
-    private void processClient(Pair<Long, Long> serverVersionPair) {
-        try {
-            ECSyncHelper ecUpdater = ECSyncHelper.getInstance(context);
-            List<EventClient> events = ecUpdater.allEventClients(serverVersionPair.first - 1, serverVersionPair.second);
-            DrishtiApplication.getInstance().getClientProcessor().processClient(events);
-            sendSyncStatusBroadcastMessage(FetchStatus.fetched);
-        } catch (Exception e) {
-            Timber.e(e, "Process Client Exception: %s", e.getMessage());
-        }
-    }
-
-    // PUSH TO SERVER
-    private void pushToServer() {
-        pushECToServer();
-    }
-
-    private void pushECToServer() {
-        EventClientRepository db = CoreLibrary.getInstance().context().getEventClientRepository();
-        boolean keepSyncing = true;
-
-        while (keepSyncing) {
-            Map<String, Object> pendingEvents = db.getUnSyncedEvents(EVENT_PUSH_LIMIT);
-
-            if (pendingEvents.isEmpty()) {
-                return;
-            }
-
-            String baseUrl = CoreLibrary.getInstance().context().configuration().dristhiBaseURL();
-            if (baseUrl.endsWith(context.getString(R.string.url_separator))) {
-                baseUrl = baseUrl.substring(0, baseUrl.lastIndexOf(context.getString(R.string.url_separator)));
-            }
-            // create request body
-            JSONObject request = new JSONObject();
-            try {
-                if (pendingEvents.containsKey(AllConstants.KEY.CLIENTS)) {
-                    request.put(AllConstants.KEY.CLIENTS, pendingEvents.get(AllConstants.KEY.CLIENTS));
-                }
-                if (pendingEvents.containsKey(AllConstants.KEY.EVENTS)) {
-                    request.put(AllConstants.KEY.EVENTS, pendingEvents.get(AllConstants.KEY.EVENTS));
-                }
-            } catch (JSONException e) {
-                Timber.e(e);
-            }
-            String jsonPayload = request.toString();
-            Response<String> response = httpAgent.post(
-                    MessageFormat.format("{0}/{1}",
-                            baseUrl,
-                            ADD_URL),
-                    jsonPayload);
-            if (response.isFailure()) {
-                Timber.e("Events sync failed.");
-                return;
-            }
-            db.markEventsAsSynced(pendingEvents);
-            Timber.i("Events synced successfully.");
-        }
-    }
-
-    private void sendSyncStatusBroadcastMessage(FetchStatus fetchStatus) {
-        Intent intent = new Intent();
-        intent.setAction(SyncStatusBroadcastReceiver.ACTION_SYNC_STATUS);
-        intent.putExtra(SyncStatusBroadcastReceiver.EXTRA_FETCH_STATUS, fetchStatus);
-        sendBroadcast(intent);
-    }
-
-    private void complete(FetchStatus fetchStatus) {
-        Intent intent = new Intent();
-        intent.setAction(SyncStatusBroadcastReceiver.ACTION_SYNC_STATUS);
-        intent.putExtra(SyncStatusBroadcastReceiver.EXTRA_FETCH_STATUS, fetchStatus);
-        intent.putExtra(SyncStatusBroadcastReceiver.EXTRA_COMPLETE_STATUS, true);
-
-        sendBroadcast(intent);
-        //sync time not update if sync is fail
-        if (!fetchStatus.equals(FetchStatus.noConnection) && !fetchStatus.equals(FetchStatus.fetchedFailed)) {
-            ECSyncHelper ecSyncUpdater = ECSyncHelper.getInstance(context);
-            ecSyncUpdater.updateLastCheckTimeStamp(new Date().getTime());
-        }
-
-    }
-
-    private Pair<Long, Long> getMinMaxServerVersions(JSONObject jsonObject) {
-        final String EVENTS = "events";
-        final String SERVER_VERSION = "serverVersion";
-        try {
-            if (jsonObject != null && jsonObject.has(EVENTS)) {
-                JSONArray events = jsonObject.getJSONArray(EVENTS);
-
-                long maxServerVersion = Long.MIN_VALUE;
-                long minServerVersion = Long.MAX_VALUE;
-
-                for (int i = 0; i < events.length(); i++) {
-                    Object o = events.get(i);
-                    if (o instanceof JSONObject) {
-                        JSONObject jo = (JSONObject) o;
-                        if (jo.has(SERVER_VERSION)) {
-                            long serverVersion = jo.getLong(SERVER_VERSION);
-                            if (serverVersion > maxServerVersion) {
-                                maxServerVersion = serverVersion;
-                            }
-
-                            if (serverVersion < minServerVersion) {
-                                minServerVersion = serverVersion;
-                            }
-                        }
-                    }
-                }
-                return Pair.create(minServerVersion, maxServerVersion);
-            }
-        } catch (Exception e) {
-            Timber.e(e);
-        }
-        return Pair.create(0L, 0L);
-    }
-
-    private int fetchNumberOfEvents(JSONObject jsonObject) {
-        int count = -1;
-        final String NO_OF_EVENTS = "no_of_events";
-        try {
-            if (jsonObject != null && jsonObject.has(NO_OF_EVENTS)) {
-                count = jsonObject.getInt(NO_OF_EVENTS);
-            }
-        } catch (JSONException e) {
-            Timber.e(e);
-        }
-        return count;
-    }
-
-    public int getEventPullLimit() {
-        return EVENT_PULL_LIMIT;
-    }
-
-    private synchronized void fetchMissingEventsRetry(final int count, List<Task> tasksWithMissingClientsEvents) {
-        try {
-            final ECSyncHelper ecSyncUpdater = ECSyncHelper.getInstance(context);
-            if (httpAgent == null) {
-                complete(FetchStatus.fetchedFailed);
-            }
-
-            String baseUrl = CoreLibrary.getInstance().context().
-                    configuration().dristhiBaseURL();
-            if (baseUrl.endsWith("/")) {
-                baseUrl = baseUrl.substring(0, baseUrl.lastIndexOf("/"));
-            }
-
-            if (tasksWithMissingClientsEvents.isEmpty()) {
-                return;
-            }
-
-            List<String> missingEventIdsInTasks = new ArrayList<>();
-            StringBuilder baseEntityIdsSb = new StringBuilder();
-            for (int i = 0; i < tasksWithMissingClientsEvents.size(); i++) {
-                Task task = tasksWithMissingClientsEvents.get(i);
-                baseEntityIdsSb.append(task.getForEntity());
-                missingEventIdsInTasks.add(task.getReasonReference());
-                // if not the last item
-                if (i != tasksWithMissingClientsEvents.size() - 1) {
-                    baseEntityIdsSb.append(",");
-                }
-
-            }
-
-            SyncConfiguration configs = CoreLibrary.getInstance().getSyncConfiguration();
-            String url = baseUrl + SYNC_URL;
-            Response resp;
-            if (configs.isSyncUsingPost()) {
-                JSONObject syncParams = new JSONObject();
-                syncParams.put("baseEntityId", baseEntityIdsSb.toString());
-                syncParams.put("serverVersion", 0);
-                syncParams.put("limit", EVENT_PULL_LIMIT);
-                resp = httpAgent.postWithJsonResponse(url, syncParams.toString());
-            } else {
-                url += "?baseEntityId=" + baseEntityIdsSb.toString() + "&serverVersion=0&limit=" + EVENT_PULL_LIMIT;
-                Timber.i("URL: %s", url);
-                resp = httpAgent.fetch(url);
-            }
-
-            if (resp.isUrlError()) {
-                FetchStatus.fetchedFailed.setDisplayValue(resp.status().displayValue());
-                complete(FetchStatus.fetchedFailed);
-                return;
-            }
-
-            if (resp.isTimeoutError()) {
-                FetchStatus.fetchedFailed.setDisplayValue(resp.status().displayValue());
-                complete(FetchStatus.fetchedFailed);
-            }
-
-            if (resp.isFailure() && !resp.isUrlError() && !resp.isTimeoutError()) {
-                fetchMissingEventsFailed(count, tasksWithMissingClientsEvents);
-            }
-
-            int eCount;
-            JSONObject jsonObject = new JSONObject();
-            if (resp.payload() == null) {
-                eCount = 0;
-            } else {
-                jsonObject = new JSONObject((String) resp.payload());
-                eCount = fetchNumberOfEvents(jsonObject);
-                Timber.i("Parse Network Event Count: %s", eCount);
-            }
-
-            if (eCount < 0) {
-                fetchMissingEventsFailed(count, tasksWithMissingClientsEvents);
-            } else {
-                final Pair<Long, Long> serverVersionPair = getMinMaxServerVersions(jsonObject);
-                filterEventsForOnlyMissingEventsInTasks(jsonObject, missingEventIdsInTasks);
-                boolean isSaved = ecSyncUpdater.saveAllClientsAndEvents(jsonObject);
-                if (isSaved) {
-                    processClient(serverVersionPair);
-                }
-            }
-        } catch (Exception e) {
-            Timber.e(e, "Fetch Retry Exception:  %s", e.getMessage());
-            fetchMissingEventsFailed(count, tasksWithMissingClientsEvents);
-        }
-    }
-
-    public void filterEventsForOnlyMissingEventsInTasks(JSONObject jsonObject, List<String> missingEventIdsInTasks) {
-        JSONArray missingEvents = new JSONArray();
-        try {
-            JSONArray receivedeEvents = jsonObject.getJSONArray("events");
-            for (int i = 0; i < receivedeEvents.length(); i++) {
-                JSONObject event = receivedeEvents.getJSONObject(i);
-                if (missingEventIdsInTasks.contains(event.getString("id"))) {
-                    missingEvents.put(event);
-                }
-            }
-            jsonObject.put("no_of_events", missingEvents.length());
-            jsonObject.put("events", missingEvents);
-        } catch (Exception e) {
-            Timber.e(e);
-        }
-    }
-
-    public void fetchMissingEventsFailed(int count, List<Task> tasksWithMissingClientsEvents) {
-        if (count < CoreLibrary.getInstance().getSyncConfiguration().getSyncMaxRetries()) {
-            int newCount = count + 1;
-            fetchMissingEventsRetry(newCount, tasksWithMissingClientsEvents);
-        } else {
-            complete(FetchStatus.fetchedFailed);
-        }
-    }
 }
